% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialGen.R
\name{materialGeneratorUI}
\alias{materialGeneratorUI}
\title{物料编码生成器界面}
\usage{
materialGeneratorUI(
  tabTitle = "物料编码生成器",
  colTitles = c("物料属性设置", "物料生成结果"),
  widthRates = c(3, 9),
  func_left = materialGeneratorUI_com_left,
  func_right = materialGeneratorUI_com_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}
}
\value{
返回值
}
\description{
物料编码生成器界面
}
\examples{
materialGeneratorUI()
}
