% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialGen.R
\name{materialGeneratorUI_com_left}
\alias{materialGeneratorUI_com_left}
\title{请输入文件}
\usage{
materialGeneratorUI_com_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
materialGeneratorUI_com_left()
}
