% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialGen.R
\name{materialGeneratorUI_com_right}
\alias{materialGeneratorUI_com_right}
\title{生成区域}
\usage{
materialGeneratorUI_com_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
materialGeneratorUI_com_right()
}
