% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prdCategory_ParentCategory_query}
\alias{prdCategory_ParentCategory_query}
\title{请查收上级产品大类}
\usage{
prdCategory_ParentCategory_query(conn = tsda::conn_rds("cprds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
请查收上级产品大类
}
\examples{
prdCategory_ParentCategory_query()
}
