% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdGroup.R
\name{prdGroupUI}
\alias{prdGroupUI}
\title{产品分组UI}
\usage{
prdGroupUI(
  tabTitle = "产品分组",
  colTitles = c("操作区域", "报表区域"),
  widthRates = c(4, 8),
  func_left = prdGroupFunc_left,
  func_right = prdGroupFunc_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{com_left}{左}

\item{com_right}{右}
}
\value{
返回值
}
\description{
产品分组UI
}
\examples{
prdGroupUI()
}
