% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdSel.R
\name{prdSelUI}
\alias{prdSelUI}
\title{物料编码生成器界面}
\usage{
prdSelUI(
  tabTitle = "物料列表查询",
  colTitles = c("操作设置", "报表区域"),
  widthRates = c(4, 8),
  func_left = prdSelUI_com_left,
  func_right = prdSelUI_com_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}
}
\value{
返回值
}
\description{
物料编码生成器界面
}
\examples{
prdGenUI()
}
