% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propCategoryConfig.R
\name{propCategoryConfigFunc_right}
\alias{propCategoryConfigFunc_right}
\title{请输入文件}
\usage{
propCategoryConfigFunc_right()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
propCategoryConfigFunc_right()
}
