% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propCategoryConfig.R
\name{propCategoryConfigUI}
\alias{propCategoryConfigUI}
\title{产品大类UI}
\usage{
propCategoryConfigUI(
  tabTitle = "属性类别配置",
  colTitles = c("操作区域", "报表区域"),
  widthRates = c(4, 8),
  func_left = propCategoryConfigFunc_left,
  func_right = propCategoryConfigFunc_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{com_left}{左}

\item{com_right}{右}
}
\value{
返回值
}
\description{
产品大类UI
}
\examples{
propCategoryConfigUI()
}
