% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propValueConfig.R
\name{propValueConfigUI}
\alias{propValueConfigUI}
\title{产品大类UI}
\usage{
propValueConfigUI(
  tabTitle = "属性选项配置",
  colTitles = c("操作区域", "报表区域"),
  widthRates = c(4, 8),
  func_left = propValueConfigFunc_left,
  func_right = propValueConfigFunc_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{com_left}{左}

\item{com_right}{右}
}
\value{
返回值
}
\description{
产品大类UI
}
\examples{
propValueConfigUI()
}
