% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propValue.R
\name{propValueUI}
\alias{propValueUI}
\title{产品大类UI}
\usage{
propValueUI(
  tabTitle = "属性选项",
  colTitles = c("操作区域", "报表区域"),
  widthRates = c(4, 8),
  func_left = propValueFunc_left,
  func_right = propValueFunc_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{com_left}{左}

\item{com_right}{右}
}
\value{
返回值
}
\description{
产品大类UI
}
\examples{
propValueUI()
}
